Refacored user_messages; integrated anon_user.py
